Work around routing String/Integer weirdness.
Get working version of some debugger gems since ruby 1.9.3
is broken with the stable version of ruby-debug.

Remove defaults from the root routes since there's
something weird there that causes the url_for
method to not find the route if one of the default parameters
is a string and you pass in an Integer or vice-versa. Or something
like that. Fixed by handling default arguments in the controller
instead, which is fortunately very easy.

Test that page parameter handling is robust.