Add config hash for picking layout per domain.
First, the cache is checked, and if the cache contains an entry then
it uses that. Otherwise, it matches the regex keys against the domain
to figure out the layout and saves the mapping of domain to layout
in the cache hash.

The code to do this will be written soon and should apply to
all non-administrative pages. Maybe administrative pages in the
future too?